+++ Model/config_majorModel.cs
@@ -9,7 +9,7 @@ namespace Model
 {
     public class config_majorModel
     {
-        public System.Int16 Id { get; set; }
+        public int Id { get; set; }
         public System.String major_kind_id { get; set; }
         public System.String major_kind_name { get; set; }
         public System.String major_id { get; set; }
